requirements.txt
Django==5.0.3
django-environ==0.11.2
python-dotenv==1.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
ccxt==4.2.59
GoogleNews==1.6.13
pandas==2.2.1
msgpack==1.0.8
websockets==12.0
sqlalchemy==2.0.27
channels[daphne]

# POSTGRES
#If running on Linux, the below commands must be entered first
# sudo apt-get install python-pip
# sudo apt-get update
# sudo apt-get install libpq-dev python2-dev python2 python-dev-is-python3
psycopg2==2.9.9

# TA-Lib Installation:
# curl -SLO http://downloads.sourceforge.net/project/ta-lib/ta-lib/0.4.0/ta-lib-0.4.0-src.tar.gz
# tar -xzf ta-lib-0.4.0-src.tar.gz
# cd ta-lib/
# ./configure --prefix=/usr
# make
# sudo make install
TA-Lib==0.4.28
pandas-ta==0.3.14b